Configure the development database
For development, I'm using a SQLite database, which is created in the
repo's "db" directory. Other users may want to use a different DB
system, and once we deploy we'll have to have server-specific settings,
so eventually this will likely be moved out of settings.py, but for
now this will help jumpstart development.