Attach social networking fields to user profiles
Included is a demonstration of how to use them in templates.

I'm thinking about redoing this structure. Instead of separate
Facebook, Tumblr, and Twitter fields, we could just have a generic
many-to-many field to a "Service" object, which would allow us to easily
add services in the future without rewriting any code or changing the
database structure.